Fix gun powerup effects lagging behind the guns in first person
there were 2 if's checking (!cg.hyperspace), someone decided to clean this up by combining them. This resulted in trap->FX_AddScheduledEffects(qfalse); being called before CG_AddViewWeapon( &cg.predictedPlayerState ); causing effects for the first person view to not be scheduled on the correct frame, causing the effect to not be visible, later someone set the lifetime of the particle to be longer, causing it to be visible but trailing after.

this cleanup was well intended but caused a hard to spot bug if you didn't specifically test for first person guns with charging shots.

You could argue that you dont really need the if (!cg.hyperspace) here at all, but i decided to revert it to it's original working state.